cldRun/requirements.txt
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.6
Flask==2.0.1
google-api-core==2.0.1
google-auth==2.2.1
google-cloud-core==2.0.0
google-cloud-storage==1.42.2
google-crc32c==1.2.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
gunicorn==20.1.0
idna==3.2
numpy==1.21.2
pandas==1.3.3
pip==20.1.1
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
rsa==4.7.2
setuptools==47.1.0
six==1.16.0
urllib3==1.26.7
fsspec
gcsfs